refactor: modularize DynamoDB adapter operations
- Break up monolithic 446-line adapter file into modular components
- Create organized directory structure for operations:
  - operations/single: single record operations (create, fetch, delete, replace, patch)
  - operations/batch: batch operations (create-many, fetch-many, delete-many, patch-many)
  - operations/query: query operations (fetch-all, create-fetch-all)
- Extract all 11 operations into separate focused modules
- Create factory module to assemble adapter from modular operations
- Add constants file for DynamoDB limits
- Simplify main adapter file to re-export from factory
- Maintain backward compatibility - all tests passing

This refactoring improves maintainability, testability, and follows SOLID principles
by separating concerns into focused, single-responsibility modules.
