+++ src/adapters/dynamodb/operations/batch/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Batch operations for DynamoDB
+ */
+
+export { createCreateManyRecords } from './create-many.operation';
+export { createFetchManyRecords } from './fetch-many.operation';
+export { createDeleteManyRecords } from './delete-many.operation';
+export { createPatchManyRecords } from './patch-many.operation';
