+++ src/adapters/dynamodb/operations/query/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Query operations for DynamoDB
+ */
+
+export { createFetchAllRecords } from './fetch-all.operation';
+export { createCreateFetchAllRecords } from './create-fetch-all.operation';
