+++ src/adapters/dynamodb/operations/single/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Single item operations for DynamoDB
+ */
+
+export { createCreateOneRecord } from './create.operation';
+export { createFetchOneRecord } from './fetch.operation';
+export { createDeleteOneRecord } from './delete.operation';
+export { createReplaceOneRecord } from './replace.operation';
+export { createPatchOneRecord } from './patch.operation';
